Generalise scan command
The scan command now takes a single kubernetes resource file as argument
unless a switch is specified to say that the input file is a kab manifest.

The output is printed to standard output unless and output file is specified.

"Scanning ..." messages are printed to standard error for ease of piping
standard output.